backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
slowapi>=0.1.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 